ADC10 now seems to work in debug setup.
ADC is used in multiple channel read mode. All three IR diode voltages are read in single run. This could as well be implemented for all available analog signals (FOOD, BATTERY, TEMPERATURE). Velues are not verified yet. Also the currently used 2.5V reference might not be the best choice, as the IR diods are driven by Vbatt which varies over time. Vbatt most likely will be the best choice.

TODO:
* select optimal reference voltage
* verify that the values are correct and correspond to the light situation.
* implement all analog values

+++ adc10.h
@@ -27,7 +27,7 @@ void adc10_init( void );
 void adc10_grab( void );
 
 /* The readings from the photodiodes */
-extern uint16_t pd_value[3];
+extern int pd_value[3];
 
 /* Which photodiode is which */
 #define PD_FRONT 0
